Add files via upload
This is the vending_machine.py file that works with windows (Servo Motor commented out). The accompanying test_vending_machine.py file test the state machine coin values by ensuring the correct amount (in cents) and that the vending machine cycles to the correct state with regards to depositing coins.